refactor(examples): tighten the session-handoff demo without changing a byte
The racing-source format string now lives in one place and the handoff act
imports it, so the expectation and the writer cannot drift apart silently. The
runner and tests compare restore outcomes against the RESTORE_OUTCOME constants
rather than raw strings, as the workspace-versioning demo already does, and the
test fixture builds its workspace through the same helper the acts use, which
also stops the coordinator's tighten warning firing under pytest.

All three act entry points constructed session A before the try whose finally
removes the temp workspace, so a failed spawn orphaned the directory; A is now
built inside it and closed only if it exists. The child no longer formats a
traceback for a deny the parent never reads, and the runner's redundant path
bootstrap — dead because the package init always runs first — is gone.

Stdout across a run is byte-identical to the previous commit and stderr stays
empty.
